requirements.txt
flask
bcrypt
pyodbc
mysql-connector-python